Fix u32 return local in FUN_0036e2f0

+++ src/Event/mt_evtMain.c
@@ -14541,7 +14541,7 @@ u32 FUN_0036e2f0(int param_1)
 
   long lVar9;
 
-  u64 uVar10;
+  u32 uVar10;
 
   u32 *puVar11;
